[cpu][arm]: create arm namespace
more and more functions had prefix arm_
it was a sign that it is time to create a seperate namespace.

added 2 types: arm::instruction and thumb::instruction.
it just nice to refer to instruction just as an instruction regardless
of whether it is thumb or arm.
what kind of instruction it is, can be infered from an enclosing
namespace.